pyproject.toml
[tool.poetry]
name = "flaskurlshortener"
version = "0.1.0"
description = ""
authors = ["advait-panicker <advait.panicker@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
pymongo = "^4.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
